bug fix api

+++ app/Http/Controllers/Api/V1/CricketDetail/Match/MatchMastersControllerApi.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Api\V1\CricketDetail\Match;
 
 use Illuminate\Http\Request;
 use Auth;
-
+use Validator;
 use App\Model\MatchMaster_model;
 use App\Model\UserOrganisation_model;
 use App\Model\UserMaster_model;
@@ -41,48 +41,56 @@ class MatchMastersControllerApi extends Controller
     }
 
     public function listMatch(Request $request){
-        $this->validate($request,[
+        $validator = Validator::make($request->all(), [
             'tournament_id' => 'required|numeric|min:1',
             'organization_master_id' => 'required|numeric|min:1',
         ]);
-        $organization_master_id = 1;//have to find Org id from login
-        $Tour_id = $this->TournamentMaster_model->getId($organization_master_id,$request->tournament_id);
-        $Matches = $this->MatchMaster_model->checkTourId($Tour_id);
-        if($Matches){
-            $output = array('status' => 200 ,'msg' => 'Sucess','data' => $Matches);
-        }else{
-            $output = array('status' => 400 ,'msg' => 'Transection Fail');
-        }
-        return response()->json($output);
-    }
-    
-    public function addMatch(Request $request){
-        $this->validate($request, [
-            'tournament_id' => 'required|numeric',
-            'team1' => 'required|numeric',
-            'team2' => 'required|numeric',
-            'match_name' => 'required|max:190',
-            'ground_name' => 'required|max:190',
-            'match_type' => 'required|max:190',
-            'match_date' => 'required|date|after:'.date('Y-m-d'),
-            'overs' => 'required|numeric',
-            'innings' => 'required|numeric',
-        ]);
-        if(!$request->team1 == $request->team2){
-            $Matches = $this->MatchMaster_model->SaveMatch($request);
+        if(!$validator->fails()){
+            $organization_master_id = 1;//have to find Org id from login
+            $Tour_id = $this->TournamentMaster_model->getId($organization_master_id,$request->tournament_id);
+            $Matches = $this->MatchMaster_model->checkTourId($Tour_id);
             if($Matches){
                 $output = array('status' => 200 ,'msg' => 'Sucess','data' => $Matches);
             }else{
                 $output = array('status' => 400 ,'msg' => 'Transection Fail');
             }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Please Select Another Team');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
+        }
+        return response()->json($output);
+    }
+    
+    public function addMatch(Request $request){
+        $validator = Validator::make($request->all(), [
+            'tournament_id' => 'required|numeric',
+            'team1' => 'required|numeric',
+            'team2' => 'required|numeric',
+            'match_name' => 'required|max:190',
+            'ground_name' => 'required|max:190',
+            'match_type' => 'required|max:190',
+            'match_date' => 'required|date|after:'.date('Y-m-d'),
+            'overs' => 'required|numeric',
+            'innings' => 'required|numeric',
+        ]);
+        if(!$validator->fails()){
+            if(!$request->team1 == $request->team2){
+                $Matches = $this->MatchMaster_model->SaveMatch($request);
+                if($Matches){
+                    $output = array('status' => 200 ,'msg' => 'Sucess','data' => $Matches);
+                }else{
+                    $output = array('status' => 400 ,'msg' => 'Transection Fail');
+                }
+            }else{
+                $output = array('status' => 400 ,'msg' => 'Please Select Another Team');
+            }
+        }else{
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
 
     public function updateMatch(Request $request){
-        $this->validate($request, [
+        $validator = Validator::make($request->all(), [
             'tournament_id' => 'required|numeric',
             'match_id' => 'required|numeric',
             'team1' => 'required|numeric',
@@ -94,31 +102,39 @@ class MatchMastersControllerApi extends Controller
             'overs' => 'required|numeric',
             'innings' => 'required|numeric',
         ]);
-        if(!$request->team1 == $request->team2){
-            $Match = $this->MatchMaster_model->getDetailByTourMatch($request->tournament_id,$request->match_id);
-            if($Match){
-                $this->MatchMaster_model->updateByTourId($request->tournament_id,$request->match_id,$request);
-                $output = array('status' => 200 ,'msg' => 'Sucess');
+        if(!$validator->fails()){
+            if(!$request->team1 == $request->team2){
+                $Match = $this->MatchMaster_model->getDetailByTourMatch($request->tournament_id,$request->match_id);
+                if($Match){
+                    $this->MatchMaster_model->updateByTourId($request->tournament_id,$request->match_id,$request);
+                    $output = array('status' => 200 ,'msg' => 'Sucess');
+                }else{
+                    $output = array('status' => 400 ,'msg' => 'Transection Fail');
+                }
             }else{
-                $output = array('status' => 400 ,'msg' => 'Transection Fail');
+                $output = array('status' => 400 ,'msg' => 'Please Select Another Team');
             }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Please Select Another Team');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
 
     public function deleteMatch(Request $request){
-        $this->validate($request, [
+        $validator = Validator::make($request->all(), [
             'tournament_id' => 'required|numeric',
             'match_id' => 'required|numeric',
         ]);
-        $Match = $this->MatchMaster_model->getDetailByTourMatch($request->tournament_id,$request->match_id);
-        if($Match){
-            $Match = $this->MatchMaster_model->deleteByTourMatch($request->tournament_id,$request->match_id);
-            $output = array('status' => 200 ,'msg' => 'Sucess');
+        if(!$validator->fails()){
+            $Match = $this->MatchMaster_model->getDetailByTourMatch($request->tournament_id,$request->match_id);
+            if($Match){
+                $Match = $this->MatchMaster_model->deleteByTourMatch($request->tournament_id,$request->match_id);
+                $output = array('status' => 200 ,'msg' => 'Sucess');
+            }else{
+                $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
