+++ app/Http/Controllers/Api/V1/Users/Player/UserAchievementControllerApi.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Api\V1\Users\Player;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Auth;
-
+use Validator;
 use App\Model\UserMaster_model;
 use App\Model\UserAchievement_model;
 use App\Model\OrganisationMaster_model;
@@ -38,26 +38,30 @@ class UserAchievementControllerApi extends Controller{
     }
 
     public function addAchievement(Request $request){
-        $this->validate($request,[
+        $validator = Validator::make($request->all(), [
             'title' => 'required|max:190',
             'location' => 'required|max:190',
             '$organization_master_id' => 'required|numeric|min:1',
             'start_date' => 'required|date|before:end_date',
             'end_date' => 'required|date|after:start_date',
         ]);
-        $user_master_id = 1; //have to find
-        $request->request->add(['user_master_id' => $user_master_id]);
-        $User_Achieve = $this->UserAchievement_model->SaveAchievement($request);
-        if($User_Achieve){
-            $output = array('status' => 200 ,'msg' => 'Sucess');
+        if(!$validator->fails()){
+            $user_master_id = 1; //have to find
+            $request->request->add(['user_master_id' => $user_master_id]);
+            $User_Achieve = $this->UserAchievement_model->SaveAchievement($request);
+            if($User_Achieve){
+                $output = array('status' => 200 ,'msg' => 'Sucess');
+            }else{
+                $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
 
     public function updateAchievement(Request $request) {
-        $this->validate($request,[
+        $validator = Validator::make($request->all(), [
             'id' => 'required|numeric|min:1',
             '$organization_master_id' => 'required|numeric|min:1',
             'title' => 'required|max:190',
@@ -65,27 +69,35 @@ class UserAchievementControllerApi extends Controller{
             'start_date' => 'required|date|before:end_date',
             'end_date' => 'required|date|after:start_date',
         ]);
-        $user_master_id = 1; //have to find
-        $request->request->add(['user_master_id' => $user_master_id,'update' => 1,'id' => $request->id]);
-        $User_Achieve = $this->UserAchievement_model->SaveUserBio($request);
-        if($User_Achieve){
-            $output = array('status' => 200 ,'msg' => 'Sucess');
+        if(!$validator->fails()){
+            $user_master_id = 1; //have to find
+            $request->request->add(['user_master_id' => $user_master_id,'update' => 1,'id' => $request->id]);
+            $User_Achieve = $this->UserAchievement_model->SaveUserBio($request);
+            if($User_Achieve){
+                $output = array('status' => 200 ,'msg' => 'Sucess');
+            }else{
+                $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
 
     public function deleteAchievement(Request $request){
-        $this->validate($request,[
+        $validator = Validator::make($request->all(), [
             'id' => 'required|numeric|min:1',
         ]);
-        $User_Achieve = $this->UserAchievement_model->getById($request->id);
-        if($User_Achieve){
-            $User_Achieve->delete();
-            $output = array('status' => 200 ,'msg' => 'Sucess');
+        if(!$validator->fails()){
+            $User_Achieve = $this->UserAchievement_model->getById($request->id);
+            if($User_Achieve){
+                $User_Achieve->delete();
+                $output = array('status' => 200 ,'msg' => 'Sucess');
+            }else{
+                $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            }
         }else{
-            $output = array('status' => 400 ,'msg' => 'Transection Fail');
+            $output = array('status' => 400 ,'msg' => 'Transection Fail','errors' => $validator->errors()->all());
         }
         return response()->json($output);
     }
